% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol consumption}
\format{
A data frame with 193 rows and 5 variables.
\describe{
\item{country}{Country name}
\item{beer_servings}{beer servings of the country}
\item{spirit_servings}{spirit servings of the country}
\item{wine_servings}{wine servings of the country}
\item{total_litres_of_pure_alcohol}{the total (sum of recorded and unrecorded alcohol) amount of alcohol consumed per person (15 years of age or older) over a calendar year, in litres of pure alcohol, adjusted for tourist consumption. (The World Bank Group,2022)}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-06-26/week13_alcohol_global.csv}
}
\usage{
alcohol
}
\description{
A dataset that contains beer, spirit and wine servings and total litres of pure alcohol by country.
}
\keyword{datasets}
