% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdata.R
\docType{package}
\name{asdata}
\alias{asdata}
\title{asdata: Global alcohol consumption and servings}
\description{
A package gives a good overview of national alcohol consumption and the servings of beer, spirit and wine,
and makes it easy to compare alcohol consumption and servings from country to country.
}
\section{Authors}{


{Di Cui \email{dcui0002@student.monash.edu}}
}

\section{Functions}{

The asdata package provides 5 important functions:
\itemize{
\item{\code{\link[=calculate_country_alcohol]{calculate_country_alcohol()}}}
\item{\code{\link[=extract_country_servings]{extract_country_servings()}}}
\item{\code{\link[=plot_servings]{plot_servings()}}}
\item{\code{\link[=plotly_litres]{plotly_litres()}}}
\item{\code{\link[=run_app]{run_app()}}}
}

The asdata package provides 2 external data objects:
\itemize{
\item{\link{alcohol_map}}
\item{\link{alcohol}}
}
}

