% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_country_alcohol.R
\name{calculate_country_alcohol}
\alias{calculate_country_alcohol}
\title{Difference in global alcohol consumption and servings}
\usage{
calculate_country_alcohol(name, type)
}
\arguments{
\item{name}{Vector.Vector of country names.}

\item{type}{Character. The type of alcohol value, one of "beer_servings",
"spirit_servings", "wine_servings" and "total_litres_of_pure_alcohol".}
}
\value{
A value. Mean of alcohol value of selected country.

A data.frame. Difference alcohol value between country.
\itemize{
\item \code{country} : Country name.
\item \code{diff} : Difference between alcohol value of country and average alcohol value of selected country.
\item \code{beer_servings/ spirit_servings/ wine_servings/ total_litres_of_pure_alcohol} : value of beer_servings, spirit_servings, wine_servings or total_litres_of_pure_alcohol.
}
}
\description{
This function calculates the country average alcohol consumption
and servings, as well as the difference between the average and the country.
}
\examples{
# Average wine servings of China, Japan, France, Australia and USA,
# a dataset of difference between wine_servings of each country with the average wine_servings.
country <- c("China","Japan","France", "Australia","USA")
calculate_country_alcohol(country, "wine_servings")


}
