% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_country_servings.R
\name{extract_country_servings}
\alias{extract_country_servings}
\title{Extract country servings}
\usage{
extract_country_servings(range, type)
}
\arguments{
\item{range}{Vector. The set range of alcohol servings
The max servings of beer is \emph{376}, the max servings of spirit is \emph{438},
and the max servings of wine is \emph{370}.}

\item{type}{Character. The type of alcohol servings, one of "beer_servings",
"spirit_servings" and "wine_servings".}
}
\value{
A data.frame. Country servings.
\itemize{
\item \code{country} : Country name.
\item \code{rank} : Ranking of the country's (beer/spirit/wine) servings.
\item \code{beer/spirit/wine servings} : (Beer/Spirit/Wine) servings of the country.
}
}
\description{
This function is used to obtain a list of countries in this
alcohol servings range in descending order of this alcohol servings.
}
\examples{
# countries with beer servings in the range (0,100)
extract_country_servings(range = c(0,100), type = "beer_servings")


}
