% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get alcohol and alcohol_map dataset}
\usage{
get_data(data)
}
\arguments{
\item{data}{Character. Either "alcohol" or "alcohol_map" to get dataset.
"alcohol" get alcohol dataset, which contains beer, spirit and
wine servings and total litres of pure alcohol by country.
"alcohol_map" get alcohol_map dataset, which contains alcohol dataset and world map.}
}
\value{
A data.frame. alcohol dataset or alcohol_map dataset.
}
\description{
This function could get alcohol dataset and alcohol_map dataset,
which contain beer, spirit and wine servings, total litres of pure alcohol
and geometry information by country.
}
\examples{
# Get alcohol dataset.
get_data(data = "alcohol")

}
