% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_litres.R
\name{plot_litres}
\alias{plot_litres}
\title{Plot global alcohol consumption map}
\usage{
plot_litres(Litres)
}
\arguments{
\item{Litres}{Vector. The total litres of pure alcohol shown in the graph is in this range.
The total litres of pure alcohol worldwide ranges from \emph{0} to \emph{14.4} liters.}
}
\value{
A map plot. Global alcohol consumption map.
}
\description{
This function plots global alcohol consumption map in a total litres of pure alcohol range.
}
\examples{
# a map plot shows global alcohol consumption with total litres of pure alcohol in range (0,10)
plot_litres(c(0,10))


}
