% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_servings.R
\name{plot_servings}
\alias{plot_servings}
\title{Mapping global servings of beer, spirit and wine}
\usage{
plot_servings(type)
}
\arguments{
\item{type}{Character. The type of alcohol servings, one of "beer_servings",
"spirit_servings" and "wine_servings".}
}
\description{
This function maps global alcohol servings of beer, spirit and wine.
}
\examples{
# maps global beer servings.
plot_servings("beer_servings")

}
