% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run shiny app}
\usage{
run_app()
}
\value{
A shiny app.
}
\description{
Run shiny app: Global Alcohol Comsumption, which shows global beer, spirit and wine servings and total litres of pure alcohol.
}
\examples{
run_app

}
