% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_country_alcohol.R
\name{calculate_country_alcohol}
\alias{calculate_country_alcohol}
\title{Difference between alcohol values and average}
\usage{
calculate_country_alcohol(name, type)
}
\arguments{
\item{name}{Character. Country name. Note: For Russia, using "Russian Federation".}

\item{type}{Character. The type of alcohol value, one of "beer_servings",
"spirit_servings", "wine_servings" and "total_litres_of_pure_alcohol".}
}
\value{
A list. A list of length 2.
A Character. Mean of alcohol value of selected country.

A data.frame. Difference alcohol value between country in descending order of difference.
\itemize{
\item \code{country} : Country name.
\item \code{diff} : Difference between alcohol value of country and average alcohol value of selected country.
\item \code{beer_servings/ spirit_servings/ wine_servings/ total_litres_of_pure_alcohol} : value of beer_servings, spirit_servings, wine_servings or total_litres_of_pure_alcohol.
}
}
\description{
function calculates the average alcohol consumption and servings of selected countries,
as well as the difference between the alcohol consumption and servings of each country  and the average.
}
\examples{
# calculate average wine servings of China, Japan, France, Australia and USA,
# Calculate difference between wine_servings of each country with the average wine_servings.
country <- c("China","Japan","France", "Australia","USA")
calculate_country_alcohol(country, "wine_servings")


}
